requirements.txt
aiodns
cchardet
aiohttp
asyncio
python-dotenv
coloredlogs